django_react2/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
coverage==7.2.7
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
